feat: move user creation logic to dashboard page; remove it from the with-navbar page
