+++ src/app/(with-navbar)/page.tsx
@@ -6,21 +6,8 @@ import Instructors from '@/components/Instructors';
 import Footer from '@/components/Footer';
 import Faq from '@/components/Faq';
 import { Globe } from '@/components/ui/globe';
-import { useEffect } from 'react';
-import axios from 'axios';
 
 export default function Page() {
-    useEffect(() => {
-        const createUser = async () => {
-            try {
-                await axios.post('/api/user');
-            } catch (error) {
-                console.error('Error creating user:', error);
-            }
-        };
-
-        createUser();
-    }, []);
     return (
         <main className="min-h-screen bg-black antialiased bg-grid-white/[0.02]">
             <HeroSection />
